+++ addOns/ascanrulesBeta/src/test/java/org/zaproxy/zap/extension/ascanrulesBeta/ActiveScannerTest.java
@@ -19,7 +19,6 @@
  */
 package org.zaproxy.zap.extension.ascanrulesBeta;
 
-import org.junit.jupiter.api.Test;
 import org.parosproxy.paros.core.scanner.AbstractPlugin;
 import org.zaproxy.zap.testutils.ActiveScannerTestUtils;
 
@@ -28,11 +27,5 @@ abstract class ActiveScannerTest<T extends AbstractPlugin> extends ActiveScanner
     @Override
     protected void setUpMessages() {
         mockMessages(new ExtensionAscanRulesBeta());
-    }
-
-    @Test
-    @Override
-    public void shouldHaveValidReferences() {
-        super.shouldHaveValidReferences();
     }
 }