+++ addOns/graaljs/src/test/java/org/zaproxy/zap/extension/graaljs/PassiveDefaultTemplateGraalJsScriptTest.java
@@ -70,7 +70,7 @@ class PassiveDefaultTemplateGraalJsScriptTest extends GraalJsPassiveScriptScanRu
         assertThat(alertsRaised, hasSize(1));
         Alert alert = alertsRaised.get(0);
         assertThat(alert.getPluginId(), is(equalTo(12345)));
-        assertThat(alert.getAlertRef(), is(equalTo("12345")));
+        assertThat(alert.getAlertRef(), is(equalTo("12345-1")));
         assertThat(alert.getName(), is(equalTo("Passive Vulnerability Title")));
         assertThat(alert.getDescription(), is(equalTo("Full description")));
         assertThat(alert.getSolution(), is(equalTo("The solution")));