+++ addOns/graaljs/CHANGELOG.md
@@ -4,7 +4,8 @@ All notable changes to this add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
-
+### Changed
+- Use example links in Active/Passive Rule templates' references.
 
 ## [0.9.0] - 2025-01-09
 ### Changed